+++ libraries/DAC53001/CHANGELOG.md
@@ -6,6 +6,20 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.2.0] - 2026-09-03
+- fix #4, fix output mode, kudos to michalcourson
+- refactor bit masking code, reduce "magic numbers"
+- add error handling, CHANNEL_ERROR + PARAMETER_ERROR
+- change return type
+  - setReference() returns bool.
+  - setCurrentRange() returns bool.
+  - setOutputMode() returns bool.
+- add bool isBusy(channel)
+- update readme.md.
+- minor edits
+
+----
+
 ## [0.1.1] - 2026-01-13
 - update GitHub actions
 - minor edits